requirements.txt
# Install Django
Django==5.0.1

# Install other packages
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.2.0
PyJWT==2.8.0
bcrypt==4.1.2
twilio==8.11.1
supabase==2.3.4
PyPDF2==3.0.1
reportlab==4.0.9
django-cors-headers==4.3.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0